Changed base node numeric atributes type from int to int64

+++ node.go
@@ -10,25 +10,25 @@ type NodeInterface interface {
 	Type() string // Returns type of node
 	Name() string // Returns name of node
 
-	Id() int    // Returns id of node
-	Level() int // Returns level of node
-	Left() int  // Returns left of node
-	Right() int // Returns right of node
+	Id() int64    // Returns id of node
+	Level() int64 // Returns level of node
+	Left() int64  // Returns left of node
+	Right() int64 // Returns right of node
 
-	SetId(int)      // Sets node id
+	SetId(int64)      // Sets node id
 	SetName(string) // Sets node name
-	SetLevel(int)   // Sets node level
-	SetLeft(int)    // Sets node left
-	SetRight(int)   // Sets node right
+	SetLevel(int64)   // Sets node level
+	SetLeft(int64)    // Sets node left
+	SetRight(int64)   // Sets node right
 }
 
 // Node represents generic node type with NodeInterface implementation
 type Node struct {
-	NodeId    int    `json:"id"`
-	NodeName  string `json:"node_name"`
-	NodeLevel int    `json:"level"`
-	NodeLeft  int    `json:"left"`
-	NodeRight int    `json:"right"`
+	NodeId    int64    `json:"id"`
+	NodeName  string   `json:"node_name"`
+	NodeLevel int64    `json:"level"`
+	NodeLeft  int64    `json:"left"`
+	NodeRight int64    `json:"right"`
 }
 
 // NewNode returns a new Node instance
@@ -45,24 +45,24 @@ func (n Node) Name() string {
 	return n.NodeName
 }
 
-func (n Node) Id() int {
+func (n Node) Id() int64 {
 	return n.NodeId
 }
 
-func (n Node) Level() int {
+func (n Node) Level() int64 {
 
 	return n.NodeLevel
 }
 
-func (n Node) Left() int {
+func (n Node) Left() int64 {
 	return n.NodeLeft
 }
 
-func (n Node) Right() int {
+func (n Node) Right() int64 {
 	return n.NodeRight
 }
 
-func (n *Node) SetId(id int) {
+func (n *Node) SetId(id int64) {
 	n.NodeId = id
 }
 
@@ -70,14 +70,14 @@ func (n *Node) SetName(name string) {
 	n.NodeName = name
 }
 
-func (n *Node) SetLevel(level int) {
+func (n *Node) SetLevel(level int64) {
 	n.NodeLevel = level
 }
 
-func (n *Node) SetLeft(left int) {
+func (n *Node) SetLeft(left int64) {
 	n.NodeLeft = left
 }
 
-func (n *Node) SetRight(right int) {
+func (n *Node) SetRight(right int64) {
 	n.NodeRight = right
 }
